Implement InvalidFrequent

+++ models/config.go
@@ -3,5 +3,5 @@ package models
 type Config struct {
 	MulticastAddress string
 	LimitMessages    int
-	InvalidFrequent  float32
+	InvalidFrequent  int
 }
